Basically I made everything work. I think. I hope.
Changed operator functions to take 2 numbers, added event listeners for all buttons. Modified operate() function to take the operator and the numberArray as arguments. originally had just an if else statement to work out which operator function to use, added for loop to use the current result and current array element for equations, instead of using variables to point at array elements -1 and -2.

Hopefully everything works!